fix: avoid overflow

+++ src/hash_map.h
@@ -14,7 +14,7 @@ typedef struct {
 
 typedef struct {
     String *key;
-    int val;
+    uint64_t val;
 } Entry;
 
 typedef struct {
